+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
  - Read only `ZipStore` behind `zip` feature
+ - Add `StorageValueIO`, a [`std::io::Read`] interface to a storage value.
 
 ### Changed
  - Relax some dependency minimum versions